requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
